// Madrasaty.Front/src/app/models/examen.ts
export class Examen {
    Id: number;
    Wording: string;
    Note: number;
    DisciplineLevelId: number;
    DisciplineId: number;
    TeacherId: number;
    StartDate: Date;
    EndDate: Date;
    Discipline: string;
    Teacher: string;
    DisciplineLevel: string;
    BeginTime: Date;
    EndTime: Date;
    Periodicity:string;
    SchoolYearPeriodicityId: number;
    SubjectId: number;
    Subject:string;
    Coefficient:number;

}